requirements.txt
Django~=5.1.6
djangorestframework~=3.15.2
celery~=5.4.0
django-celery-results==2.4.0
gunicorn==20.1.0
redis==4.5.5
psycopg2-binary==2.9.10
requests~=2.31.0
python-magic~=0.4.27
django-storages~=1.14.5
drf-yasg~=1.21.9
python-decouple~=3.8
django-model-utils~=5.0.0
django-filter~=25.1
django-debug-toolbar==5.0.1
djangorestframework_simplejwt==5.5.0
boto3==1.34.107
